fix & update

+++ apps_docker/jupyter.sh
@@ -3,16 +3,17 @@ source ../linux/init.sh
 folder_to_map=/data
 
 n=jupyter
-docker ps | grep $n
-[ $? -eq 0 ] && run "docker logs $n 2>&1 | grep token"
 
-read -p "Re-create container? (y/n) " a
-[ "$a" != 'y' ] & exit
+docker ps | grep $n
+if [ $? -eq 0 ]; then
+    run "docker logs $n 2>&1 | grep token"
+    read -p "Re-create container? (y/n) " a
+    [ "$a" != 'y' ] && exit
+fi
 
 docker stop $n 2>/dev/null; docker rm $n 2>/dev/null
 
-cmd="
-docker run --name $n \
+cmd="docker run --name $n \
     --net host \
     -v $folder_to_map:$folder_to_map -w $folder_to_map \
     -d --restart unless-stopped \
@@ -33,6 +34,3 @@ cmd="$cmd fzinfz/jupyter \
 run "$cmd"
 
 run "docker logs $n -f"
-
-
-
